requirements.txt
# Runtime dependencies.
# Upper bounds are deliberate: several of these (anthropic, fastapi) make
# breaking changes between major versions. Bump only after testing.
httpx>=0.27,<1.0
Pillow>=10.2,<12.0
pillow-heif>=0.16,<1.0
python-dotenv>=1.0,<2.0
anthropic>=0.40,<1.0
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9,<1.0
numpy>=1.26,<3.0
librosa>=0.10,<1.0
soundfile>=0.12,<1.0

# Development / testing
pytest>=8.0,<9.0
